pyproject.toml
[project]
name = "observability-assistant-router"
version = "0.1.0"
description = "Multi-agent orchestrator service that routes messages to specialized agents via semantic matching"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.2.0",
    "ag-ui-protocol>=0.1.10",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "numpy>=2.0.0",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "sentence-transformers>=4.0.0",
    "sse-starlette>=2.0.0",
    "torch>=2.0.0",
    "uvicorn[standard]>=0.40.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]

# Use CPU-only PyTorch (saves ~5GB vs CUDA version)
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
